persistent-mysql: provide all definitions through an exposed "Internal" module.
